finsh all about upload files to cloudinary

+++ server/controllers/blogs.controller.js
@@ -3,6 +3,8 @@ import asyncWrapper from "../middleware/asyncWrapper.js";
 import statusValues from "../utils/statusValues.js";
 import uploadToCloudinary from "../utils/uploadToCloudinary.js";
 import { FolderPathOfCloudinary } from "../utils/constants.js";
+import replaceToCloudinary from "../utils/replaceToCloudinary.js";
+import { deleteFromCloudinary } from "../utils/deleteFromCloudinary.js";
 
 
 
@@ -39,6 +41,7 @@ const createNewBlog = asyncWrapper(
 
 const updateBlog = asyncWrapper(
     async (req, res, next) => {
+
         const { id } = req.params;
         const updatedData = req.body;
 
@@ -46,15 +49,29 @@ const updateBlog = asyncWrapper(
             return res.status(404).json({ status: statusValues.FAIL, message: 'you must provide data' })
         }
 
+        const existedBlog = await blogModel.findById(id);
+
+        if (!existedBlog) {
+            return res.status(404).json({ status: statusValues.FAIL, message: 'blog not found' })
+        }
+
+        let coverImageUrl = existedBlog.coverImage;
+        if (req.file && req.file.buffer) {
+            const result = await replaceToCloudinary(req.file.buffer, coverImageUrl);
+            coverImageUrl = result.secure_url;
+        }
+
         const updatedBlog = await blogModel.findByIdAndUpdate(
             id,
-            updatedData,
+            { ...updatedData, coverImage: coverImageUrl },
             { new: true, runValidators: true }
         )
+
         if (!updatedBlog) {
             return res.status(404)
                 .json({ status: statusValues.FAIL, message: 'blog not found' })
         }
+
         res.status(201).json({ status: statusValues.SUCCESS, message: 'blog updated successfully' })
     }
 )
@@ -62,6 +79,15 @@ const updateBlog = asyncWrapper(
 const deleteBlog = asyncWrapper(
     async (req, res, next) => {
         const { id } = req.params;
+
+        const existedBlog = await blogModel.findById(id);
+
+        if (!existedBlog) {
+            return res.status(404).json({ status: statusValues.FAIL, message: 'blog not found' })
+        }
+
+        await deleteFromCloudinary(existedBlog.coverImage);
+        
         const deletedBlog = await blogModel.findByIdAndDelete(id);
 
         if (!deletedBlog) {
